app/requirements.txt
asgiref==3.4.1
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.0
Django==3.1.7
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
dynaconf==3.1.7
hvac==0.11.2
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.2
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rest-condition==1.0.3
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
shortuuid==1.0.8
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
zipp==3.6.0
